Convert problem6 to cmake

+++ src/problem6/problem6.cpp
@@ -3,21 +3,19 @@
 
 #include <iostream>
 
-int main()
-{
-  long int sum_of_squares = 0;
-  long int square_of_sum = 0;
-  long int difference;
-  
-  for(int i = 0; i <= 100; i++)
-  {
-    sum_of_squares += (i * i);
-    square_of_sum += i;
-  }
-  square_of_sum = square_of_sum * square_of_sum;
-  
-  difference = square_of_sum - sum_of_squares;
-  
-  std::cout << "Difference: " << difference << std::endl;
-  return 0;
+int main() {
+	size_t sum_of_squares = 0;
+	size_t square_of_sum = 0;
+	size_t difference;
+
+	for(size_t i = 0; i <= 100; i++) {
+		sum_of_squares += (i * i);
+		square_of_sum += i;
+	}
+	square_of_sum = square_of_sum * square_of_sum;
+
+	difference = square_of_sum - sum_of_squares;
+
+	std::cout << "Difference: " << difference << std::endl;
+	return 0;
 }
